Abstract PlantSelector to separate component
