+++ frontend/src/PlantSelector.js
@@ -1,20 +1,69 @@
 // eslint-disable-next-line no-unused-vars
-import React from 'react';
+import React, { useEffect, useState } from 'react';
+import { apiCall } from './utils/fetchCalls';
+import { getPlantIDTuples } from './utils/plantIDs';
 import TextField from '@material-ui/core/TextField';
 import Autocomplete from '@material-ui/lab/Autocomplete';
+import Typography from '@material-ui/core/Typography';
+import CircularProgress from '@material-ui/core/CircularProgress';
+import Box from '@material-ui/core/Box';
 
-const PlantSelector = ({plantData, choosePlant}) => {
+const PlantSelector = ({ choosePlant }) => {
+    const [plantData, setPlantData] = useState([])
+
+    useEffect(() => {
+        retrieveInitialCollection()
+    }, [])
+
+    const retrieveInitialCollection = async () => {
+        const tokenFetch = await apiCall(`http://localhost:8080/`);
+        const tokenData = await tokenFetch.json();
+        const plantList = await apiCall(`https://trefle.io/api/v1/plants?filter_not%5Bedible_part%5D=null&token=${tokenData.token}`);
+        const rawPlantData = await plantList.json();
+        const plantIDData = getPlantIDTuples(rawPlantData.data);
+        
+        if (rawPlantData.links.next) {
+            retrieveSubseqCollections(plantIDData, rawPlantData.links.next, tokenData.token)
+        } else {
+            setPlantData(plantIDData)
+        }
+    }
+
+    const retrieveSubseqCollections = async (dataSoFar, link, token) => {
+        const plantList = await apiCall(`https://trefle.io${link}&token=${token}`)
+        const rawPlantData = await plantList.json();
+        const plantIDData = getPlantIDTuples(rawPlantData.data);
+        const totalPlantData = [...plantIDData, ...dataSoFar]
+
+        if (rawPlantData.links.next) {
+            retrieveSubseqCollections(totalPlantData, rawPlantData.links.next, token)
+        } else {
+            setPlantData(totalPlantData)
+        }
+    }
+
     return (
-        <Autocomplete
-            id="plant-selector"
-            options={plantData}
-            getOptionLabel={(option) => option.name}
-            onChange={(event, plant) => {
-                choosePlant(plant);
-            }}
-            renderInput={(params) => <TextField {...params} label="Combo box" variant="outlined" />
+
+        <Box my={4}>
+            <Typography variant="h4" component="h1" gutterBottom>
+              Preact v4-beta example
+            </Typography>
+            {
+              plantData.length ? 
+                <Autocomplete
+                    id="plant-selector"
+                    options={plantData}
+                    getOptionLabel={(option) => option.name}
+                    onChange={(event, plant) => {
+                        choosePlant(plant);
+                    }}
+                    renderInput={(params) => <TextField {...params} label="Combo box" variant="outlined" />
+                    }
+                />    
+              : <CircularProgress />
             }
-        />    
+        </Box>
+
     );
 }
